Update references and move select_and_prepare.py to block_encoding/
