+++ qualtran/bloqs/block_encoding/__init__.py
@@ -1,0 +1,8 @@
+from qualtran.bloqs.block_encoding.block_encoding_base import BlockEncoding
+from qualtran.bloqs.block_encoding.chebyshev_polynomial import ChebyshevPolynomial
+from qualtran.bloqs.block_encoding.lcu_block_encoding import (
+    BlackBoxPrepare,
+    BlackBoxSelect,
+    LCUBlockEncoding,
+    LCUBlockEncodingZeroState,
+)
