+++ qualtran/bloqs/block_encoding/lcu_block_encoding.py
@@ -29,7 +29,8 @@ from qualtran import (
     Soquet,
     SoquetT,
 )
-from qualtran.bloqs.select_and_prepare import PrepareOracle, SelectOracle
+from qualtran.bloqs.block_encoding.block_encoding_base import BlockEncoding
+from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle, SelectOracle
 from qualtran.bloqs.util_bloqs import Partition
 from qualtran.symbolics import SymbolicFloat
