+++ qualtran/bloqs/state_preparation/state_preparation_alias_sampling.py
@@ -32,8 +32,8 @@ from qualtran import bloq_example, BloqDocSpec, BoundedQUInt, Register, Signatur
 from qualtran._infra.gate_with_registers import total_bits
 from qualtran.bloqs.arithmetic import LessThanEqual
 from qualtran.bloqs.basic_gates.swap import CSwap
+from qualtran.bloqs.block_encoding.lcu_select_and_prepare import PrepareOracle
 from qualtran.bloqs.data_loading.qrom import QROM
-from qualtran.bloqs.select_and_prepare import PrepareOracle
 from qualtran.bloqs.state_preparation.prepare_uniform_superposition import (
     PrepareUniformSuperposition,
 )
